Rename server/internal/pluginapi to pluginhost
The server-side pluginapi package manages plugin processes, gRPC
connections, and type conversions -- it is a host/bridge layer, not an
API definition. Renaming it to pluginhost resolves the naming collision
with sdk/pluginapi and better reflects the package's role.
