+++ server/internal/pluginhost/declarative_test.go
@@ -1,4 +1,4 @@
-package pluginapi
+package pluginhost
 
 import (
 	"context"
