+++ server/internal/pluginhost/provider_host.go
@@ -1,4 +1,4 @@
-package pluginapi
+package pluginhost
 
 import (
 	"bytes"
