+++ server/internal/pluginhost/provider_server.go
@@ -1,4 +1,4 @@
-package pluginapi
+package pluginhost
 
 import (
 	"context"
