+++ server/internal/pluginhost/runtime.go
@@ -1,4 +1,4 @@
-package pluginapi
+package pluginhost
 
 import (
 	"context"
